Rename mock files to correct naming
Files renamed.
Duplicates deleted.

Related tasks APPLINK-30588 APPLINK-30974
