+++ src/components/application_manager/test/event_engine_test.cc
@@ -35,7 +35,7 @@
 #include "application_manager/event_engine/event_observer.h"
 #include "application_manager/event_engine/event.h"
 #include "application_manager/event_engine/event_dispatcher_impl.h"
-#include "application_manager/event_observer_mock.h"
+#include "application_manager/mock_event_observer.h"
 #include "application_manager/mock_event_dispatcher.h"
 #include "smart_objects/smart_object.h"
 #include "utils/make_shared.h"
