+++ src/components/application_manager/test/include/application_manager/mock_event_observer.h
