+++ src/components/application_manager/test/state_controller/state_controller_test.cc
@@ -39,7 +39,6 @@
 #include "connection_handler/mock_connection_handler_settings.h"
 #include "connection_handler/connection_handler_impl.h"
 #include "transport_manager/mock_transport_manager.h"
-#include "application_manager/statistics_manager_mock.h"
 #include "utils/lock.h"
 #include "utils/data_accessor.h"
 #include "utils/make_shared.h"
@@ -48,6 +47,7 @@
 #include "application_manager/smart_object_keys.h"
 #include "application_manager/mock_message_helper.h"
 #include "policy/mock_policy_settings.h"
+#include "policy/usage_statistics/mock_statistics_manager.h"
 #include "protocol_handler/mock_session_observer.h"
 #include "connection_handler/mock_connection_handler.h"
 #include "application_manager/policies/mock_policy_handler_interface.h"
@@ -131,7 +131,7 @@ class StateControllerImplTest : public ::testing::Test {
       , usage_stat("0",
                    utils::SharedPtr<usage_statistics::StatisticsManager>(
 
-                       new state_controller_test::MockStatisticsManager))
+                       new usage_statistics_test::MockStatisticsManager))
       , applications_(application_set_, applications_lock_) {}
   NiceMock<application_manager_test::MockApplicationManager> app_manager_mock_;
   NiceMock<policy_test::MockPolicyHandlerInterface> policy_interface_;
