+++ src/components/include/test/connection_handler/mock_connection_handler_observer.h
@@ -40,22 +40,27 @@ namespace test {
 namespace components {
 namespace connection_handler_test {
 
-class MockConnectionHandlerObserver : public ConnectionHandlerObserver {
+class ConnectionHandlerObserverMock
+    : public ::connection_handler::ConnectionHandlerObserver {
  public:
   MOCK_METHOD1(OnDeviceListUpdated,
-      void(const connection_handler::DeviceMap& device_list));
-  MOCK_METHOD0(OnFindNewApplicationsRequest,
-      void());
+               void(const connection_handler::DeviceMap& device_list));
+  MOCK_METHOD0(OnFindNewApplicationsRequest, void());
   MOCK_METHOD1(RemoveDevice,
-      void(const connection_handler::DeviceHandle& device_handle));
+               void(const connection_handler::DeviceHandle& device_handle));
   MOCK_METHOD3(OnServiceStartedCallback,
-      bool(const connection_handler::DeviceHandle& device_handle, const int32_t& session_key, const protocol_handler::ServiceType& type));
-  MOCK_METHOD3(OnServiceEndedCallback,
-      void(const int32_t& session_key, const protocol_handler::ServiceType& type, const connection_handler::CloseSessionReason& close_reason));
-  MOCK_CONST_METHOD1(GetHandshakeContext,
+               bool(const connection_handler::DeviceHandle& device_handle,
+                    const int32_t& session_key,
+                    const protocol_handler::ServiceType& type));
+  MOCK_METHOD3(
+      OnServiceEndedCallback,
+      void(const int32_t& session_key,
+           const protocol_handler::ServiceType& type,
+           const connection_handler::CloseSessionReason& close_reason));
+  MOCK_CONST_METHOD1(
+      GetHandshakeContext,
       security_manager::SSLContext::HandshakeContext(uint32_t key));
 };
-
 
 }  // namespace connection_handler_test
 }  // namespace components
