+++ src/components/media_manager/test/include/media_manager/mock_media_adapter.h
