+++ src/components/transport_manager/test/tcp_client_listener_test.cc
@@ -31,7 +31,7 @@
  */
 
 #include "gtest/gtest.h"
-#include "transport_manager/transport_adapter_mock.h"
+#include "transport_manager/transport_adapter/mock_transport_adapter.h"
 #include "transport_manager/tcp/tcp_client_listener.h"
 #include "transport_manager/mock_transport_manager.h"
 #include "transport_manager/transport_adapter/transport_adapter_controller.h"
