+++ src/components/transport_manager/test/transport_adapter_test.cc
@@ -31,12 +31,12 @@
  */
 
 #include "gtest/gtest.h"
-#include "transport_manager/device_scanner_mock.h"
-#include "transport_manager/client_connection_listener_mock.h"
-#include "transport_manager/server_connection_factory_mock.h"
-#include "transport_manager/device_mock.h"
-#include "transport_manager/connection_mock.h"
-#include "transport_manager/transport_adapter_listener_mock.h"
+#include "transport_manager/transport_adapter/mock_device_scanner.h"
+#include "transport_manager/transport_adapter/mock_client_connection_listener.h"
+#include "transport_manager/transport_adapter/mock_server_connection_factory.h"
+#include "transport_manager/transport_adapter/mock_device.h"
+#include "transport_manager/transport_adapter/mock_connection.h"
+#include "transport_manager/transport_adapter/mock_transport_adapter_listener.h"
 #include "transport_manager/transport_adapter/mock_transport_adapter_impl.h"
 #include "transport_manager/mock_transport_manager_settings.h"
